hp gerente que no sirve ni para areglar los daños mas faciles del hp proyecto y tiene que buscar ayuda de argeol y que valla y le ayude a ese hp de camilo
